+++ batch/batch_test.go
@@ -3,9 +3,7 @@ package batch_test
 import (
 	"context"
 	"errors"
-	"fmt"
 	"math/rand"
-	"strings"
 	"sync"
 	"sync/atomic"
 	"testing"
@@ -13,120 +11,6 @@ import (
 
 	. "github.com/MasterOfBinary/gobatch/batch"
 )
-
-type testSource struct {
-	Items   []interface{}
-	Delay   time.Duration
-	WithErr error
-}
-
-func (s *testSource) Read(ctx context.Context) (<-chan interface{}, <-chan error) {
-	out := make(chan interface{})
-	errs := make(chan error, 1)
-	go func() {
-		defer close(out)
-		defer close(errs)
-		for _, item := range s.Items {
-			if s.Delay > 0 {
-				time.Sleep(s.Delay)
-			}
-			select {
-			case <-ctx.Done():
-				return
-			case out <- item:
-			}
-		}
-		if s.WithErr != nil {
-			errs <- s.WithErr
-		}
-	}()
-	return out, errs
-}
-
-type countProcessor struct {
-	count        *uint32
-	delay        time.Duration
-	processorErr error
-}
-
-func (p *countProcessor) Process(ctx context.Context, items []*Item) ([]*Item, error) {
-	if p.delay > 0 {
-		time.Sleep(p.delay)
-	}
-
-	// Make sure we only access the count pointer once and do proper nil checking
-	if p.count != nil {
-		atomic.AddUint32(p.count, uint32(len(items)))
-	}
-
-	if p.processorErr != nil {
-		return items, p.processorErr
-	}
-
-	return items, nil
-}
-
-type errorPerItemProcessor struct {
-	FailEvery int
-}
-
-func (p *errorPerItemProcessor) Process(ctx context.Context, items []*Item) ([]*Item, error) {
-	for i, item := range items {
-		if p.FailEvery > 0 && (i%p.FailEvery) == 0 {
-			item.Error = fmt.Errorf("fail item %d", item.ID)
-		}
-	}
-	return items, nil
-}
-
-// Processor that transforms item data
-type transformProcessor struct {
-	transformFn func(interface{}) interface{}
-}
-
-func (p *transformProcessor) Process(ctx context.Context, items []*Item) ([]*Item, error) {
-	select {
-	case <-ctx.Done():
-		return items, ctx.Err()
-	default:
-	}
-
-	for _, item := range items {
-		if item.Error != nil {
-			continue
-		}
-		item.Data = p.transformFn(item.Data)
-	}
-	return items, nil
-}
-
-// Processor that filters items
-type filterProcessor struct {
-	filterFn func(interface{}) bool
-}
-
-func (p *filterProcessor) Process(ctx context.Context, items []*Item) ([]*Item, error) {
-	var result []*Item
-
-	for _, item := range items {
-		select {
-		case <-ctx.Done():
-			return result, ctx.Err()
-		default:
-		}
-
-		if item.Error != nil {
-			result = append(result, item)
-			continue
-		}
-
-		if p.filterFn(item.Data) {
-			result = append(result, item)
-		}
-	}
-
-	return result, nil
-}
 
 func TestBatch_ProcessorChainingAndErrorTracking(t *testing.T) {
 	t.Run("processor chaining with individual errors", func(t *testing.T) {
@@ -640,78 +524,6 @@ func TestBatch_ConcurrentProcessing(t *testing.T) {
 	})
 }
 
-func TestBatch_DynamicConfiguration(t *testing.T) {
-	t.Run("dynamic config updates during processing", func(t *testing.T) {
-		// Start with MinItems: 50 to hold processing
-		dynamicCfg := NewDynamicConfig(&ConfigValues{
-			MinItems: 50,
-			MaxItems: 0,
-			MinTime:  0,
-			MaxTime:  0,
-		})
-
-		batch := New(dynamicCfg)
-
-		// Create items
-		const totalItems = 100
-		items := make([]interface{}, totalItems)
-		for i := 0; i < totalItems; i++ {
-			items[i] = i
-		}
-
-		src := &testSource{Items: items, Delay: 5 * time.Millisecond}
-
-		// Use separate atomic counters instead of changing the pointer
-		var beforeConfigChange uint32
-		var afterConfigChange uint32
-
-		// Use a new processor type that's safe for this test
-		proc := &testProcessor{
-			processFn: func(ctx context.Context, items []*Item) ([]*Item, error) {
-				// Add a delay to ensure the processing happens over time
-				time.Sleep(10 * time.Millisecond)
-
-				// Check if config has changed and increment appropriate counter
-				config := dynamicCfg.Get()
-				if config.MinItems == 50 {
-					atomic.AddUint32(&beforeConfigChange, uint32(len(items)))
-				} else {
-					atomic.AddUint32(&afterConfigChange, uint32(len(items)))
-				}
-
-				return items, nil
-			},
-		}
-
-		// Start batch processing with initial config
-		errs := batch.Go(context.Background(), src, proc)
-
-		// Wait a bit for some items to be read, but not processed due to MinItems: 50
-		time.Sleep(100 * time.Millisecond)
-
-		// Update config to release the items for processing
-		dynamicCfg.UpdateBatchSize(5, 10)
-
-		// Wait for completion
-		<-batch.Done()
-
-		// Drain errors
-		for range errs {
-			// Just drain
-		}
-
-		// With initial MinItems: 50, we expect no items processed initially
-		if beforeConfigChange > 0 {
-			t.Errorf("expected 0 items before config update, got %d", beforeConfigChange)
-		}
-
-		// After changing to MinItems: 5, we expect items to be processed
-		if afterConfigChange == 0 {
-			t.Error("expected items to be processed after config update, got 0")
-		}
-	})
-}
-
 func TestBatch_RobustnessAndEdgeCases(t *testing.T) {
 	t.Run("large batch handling", func(t *testing.T) {
 		// Test with a large number of items to ensure memory efficiency
@@ -826,198 +638,6 @@ func TestBatch_RobustnessAndEdgeCases(t *testing.T) {
 	})
 }
 
-func TestBatch_ErrorHandling(t *testing.T) {
-	t.Run("processor with error", func(t *testing.T) {
-		batch := New(NewConstantConfig(&ConfigValues{}))
-		src := &testSource{Items: []interface{}{1, 2, 3, 4, 5}}
-
-		procErr := errors.New("processor error")
-		proc := &countProcessor{
-			count:        new(uint32),
-			processorErr: procErr,
-		}
-
-		errs := batch.Go(context.Background(), src, proc)
-
-		var foundErr bool
-		for err := range errs {
-			if err != nil && errors.Unwrap(err) == procErr {
-				foundErr = true
-				break
-			}
-		}
-
-		<-batch.Done()
-
-		if !foundErr {
-			t.Error("expected to find processor error")
-		}
-	})
-
-	t.Run("source with error", func(t *testing.T) {
-		batch := New(NewConstantConfig(&ConfigValues{}))
-
-		srcErr := errors.New("source error")
-		src := &testSource{
-			Items:   []interface{}{1, 2, 3},
-			WithErr: srcErr,
-		}
-
-		proc := &countProcessor{count: new(uint32)}
-
-		errs := batch.Go(context.Background(), src, proc)
-
-		var foundErr bool
-		for err := range errs {
-			if err != nil && errors.Unwrap(err) == srcErr {
-				foundErr = true
-				break
-			}
-		}
-
-		<-batch.Done()
-
-		if !foundErr {
-			t.Error("expected to find source error")
-		}
-	})
-
-	t.Run("nil source handling", func(t *testing.T) {
-		batch := New(NewConstantConfig(&ConfigValues{}))
-
-		// Pass nil source
-		errs := batch.Go(context.Background(), nil)
-
-		var foundErr bool
-		var errMsg string
-		for err := range errs {
-			if err != nil {
-				foundErr = true
-				errMsg = err.Error()
-				break
-			}
-		}
-
-		<-batch.Done()
-
-		if !foundErr {
-			t.Error("expected error with nil source")
-		}
-
-		if !strings.Contains(errMsg, "source cannot be nil") {
-			t.Errorf("expected 'source cannot be nil' error, got: %s", errMsg)
-		}
-	})
-
-	t.Run("nil processor filtering", func(t *testing.T) {
-		batch := New(NewConstantConfig(&ConfigValues{}))
-		src := &testSource{Items: []interface{}{1, 2, 3}}
-
-		// Include a nil processor among valid ones
-		var count uint32
-		validProc := &countProcessor{count: &count}
-
-		// Pass a mix of nil and valid processors
-		errs := batch.Go(context.Background(), src, nil, validProc, nil)
-
-		// Count errors instead of collecting them
-		errorCount := 0
-		for range errs {
-			errorCount++
-		}
-
-		<-batch.Done()
-
-		// Valid processor should still run
-		if atomic.LoadUint32(&count) != 3 {
-			t.Errorf("expected 3 items processed, got %d", count)
-		}
-
-		// The batch should process without errors
-		if errorCount > 0 {
-			t.Errorf("expected no errors with empty processor slice, got %d errors", errorCount)
-		}
-	})
-}
-
-func TestBatch_NilChannelHandling(t *testing.T) {
-	t.Run("source returning nil output channel", func(t *testing.T) {
-		batch := New(NewConstantConfig(&ConfigValues{}))
-
-		// Create a source that returns a nil output channel
-		nilChannelSource := &nilOutputChannelSource{}
-
-		errs := batch.Go(context.Background(), nilChannelSource)
-
-		var foundErr bool
-		var errMsg string
-		for err := range errs {
-			if err != nil {
-				foundErr = true
-				errMsg = err.Error()
-				break
-			}
-		}
-
-		<-batch.Done()
-
-		if !foundErr {
-			t.Error("expected error with nil output channel")
-		}
-
-		if !strings.Contains(errMsg, "nil channel") {
-			t.Errorf("expected error about nil channels, got: %s", errMsg)
-		}
-	})
-
-	t.Run("source returning nil error channel", func(t *testing.T) {
-		batch := New(NewConstantConfig(&ConfigValues{}))
-
-		// Create a source that returns a nil error channel
-		nilChannelSource := &nilErrorChannelSource{}
-
-		errs := batch.Go(context.Background(), nilChannelSource)
-
-		var foundErr bool
-		var errMsg string
-		for err := range errs {
-			if err != nil {
-				foundErr = true
-				errMsg = err.Error()
-				break
-			}
-		}
-
-		<-batch.Done()
-
-		if !foundErr {
-			t.Error("expected error with nil error channel")
-		}
-
-		if !strings.Contains(errMsg, "nil channel") {
-			t.Errorf("expected error about nil channels, got: %s", errMsg)
-		}
-	})
-}
-
-// Source that returns a nil output channel and a valid error channel
-type nilOutputChannelSource struct{}
-
-func (s *nilOutputChannelSource) Read(ctx context.Context) (<-chan interface{}, <-chan error) {
-	errs := make(chan error)
-	close(errs)
-	return nil, errs
-}
-
-// Source that returns a valid output channel and a nil error channel
-type nilErrorChannelSource struct{}
-
-func (s *nilErrorChannelSource) Read(ctx context.Context) (<-chan interface{}, <-chan error) {
-	out := make(chan interface{})
-	close(out)
-	return out, nil
-}
-
 func TestBatch_NoProcessors(t *testing.T) {
 	t.Run("no processors provided", func(t *testing.T) {
 		batch := New(NewConstantConfig(&ConfigValues{}))
@@ -1071,16 +691,4 @@ func TestBatch_NoProcessors(t *testing.T) {
 			t.Errorf("expected no errors with empty processor slice, got %d errors", errorCount)
 		}
 	})
-}
-
-// testProcessor is a processor that applies a custom processing function
-type testProcessor struct {
-	processFn func(context.Context, []*Item) ([]*Item, error)
-}
-
-func (p *testProcessor) Process(ctx context.Context, items []*Item) ([]*Item, error) {
-	if p.processFn != nil {
-		return p.processFn(ctx, items)
-	}
-	return items, nil
 }
